sirius: Fix framework mem mgmt includes.
In LOs, the *xxhdpi* files exist in frameworks native, which are
clones of the *xxxhdpi* files in the AOSP repo. Correct the include
such that the memory management setting, e.g., the dalvik heap
size, are correctly set. Otherwise many out of memory errors occur
that result in crashes and soft reboots.

Change-Id: Iefc4e86fde6c230335225415cb61123d37093388
Signed-off-by: Alexander Diewald <Diewi@diewald-net.com>